修复RAG检索分数序列化

+++ trip-assistant/rag/retriever.py
@@ -141,4 +141,4 @@ class RAGRetriever:
         import numpy as np
         vec1 = np.array(vec1)
         vec2 = np.array(vec2)
-        return np.dot(vec1, vec2) / (np.linalg.norm(vec1) * np.linalg.norm(vec2))
+        return float(np.dot(vec1, vec2) / (np.linalg.norm(vec1) * np.linalg.norm(vec2)))
